bug 2792 : Reinstated the calls to the deprecated Before- and AfterUpdateHooks, and also to the deprecated BeforeDeletionHook. I am not entirely happy with this solution, which looks ugly.
Note that currently the Before- and AfterUpdateHooks are only called (in the auto-generated controller) for the principal record of a form, and not for any auxiliary records. It's possible that the hooks should also be called for these.

This isn't really satisfactory - it makes for ugly and prolix code - and a better solution is desirable.

More thought needed.