bug 2912 : David had
onsubmit="return trapEnterSubmissions(event, this);invokeSubmitHandlers(this);"

Because the 'return' is on the first statement, the second is not executed. If one reverses it to:

onsubmit="invokeSubmitHandlers(this);return trapEnterSubmissions(event, this);"

Then invokeSubmitHandlers(this) is called (setting up the shuffle widget handler) and also return trapEnterSubmissions(event, this); is called (trapping use of the enter key to submit the form).
status 2912 : resolved FIXED